Moving Sender to SenderBase and generalizing existing logic to add largely-untested support for notification_key-enabled sending and failover scenarios
